requirements.txt
streamlit>=1.32.0
google-genai>=1.0.0
pypdf>=4.0.0
chromadb>=0.5.0
pandas>=2.0.0
python-dotenv>=1.0.0
